requirements.txt
accelerate==1.9.0
addict==2.4.0
adjustText==1.3.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
anyio==4.3.0
arrow==1.3.0
asgiref==3.7.2
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
async-timeout==4.0.3
attrs==23.2.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backoff==2.2.1
beautifulsoup4==4.12.3
bitsandbytes==0.43.1
blessed==1.20.0
blinker==1.9.0
blis==0.7.11
boto3==1.34.49
botocore==1.34.49
cachetools==5.5.2
catalogue==2.0.10
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.21.1
cmake==3.25.0
colorama==0.4.5
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
confection==0.1.5
ConfigArgParse==1.7
conllu==6.0.0
contourpy==1.2.0
croniter==1.4.1
cycler==0.12.1
cymem==2.0.11
Cython==0.29.30
dash==3.0.0
datasets==4.0.0
dateutils==0.6.12
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
deepdiff==6.7.1
dill==0.3.8
distro==1.9.0
Django==4.2.10
dm-tree==0.1.7
docstring_parser==0.17.0
editor==1.6.6
einops==0.8.1
einops-exts==0.0.4
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
etils==0.6.0
eval_type_backport==0.2.2
exceptiongroup==1.2.0
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
fastapi==0.110.0
fastjsonschema==2.21.1
ffmpy==0.3.2
filelock==3.13.1
Flask==3.0.3
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.3.1
gast==0.4.0
gin-config==0.5.0
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.38.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.56.3
gradio==3.50.2
gradio_client==0.6.1
h11==0.14.0
hf-xet==1.1.5
hf_transfer==0.1.9
httpcore==1.0.4
httplib2==0.20.4
httpx==0.27.0
huggingface-hub==0.34.1
idna==3.6
image==1.5.33
immutabledict==2.2.1
importlib_metadata==8.6.1
importlib_resources==6.1.2
inquirer==3.2.4
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1680185408135/work
ipywidgets==8.1.5
itsdangerous==2.1.2
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.3
jiter==0.9.0
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1654730843242/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1710257447442/work
jupyterlab_widgets==3.0.13
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
langcodes==3.5.0
language_data==1.3.0
lightning==2.4.0
lightning-cloud==0.5.64
lightning-utilities==0.10.1
lit==15.0.7
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.13
narwhals==1.31.0
nbformat==5.10.4
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.2.1
nmslib==2.1.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.6.77
oauth2client==4.1.3
oauthlib==3.2.2
open-flamingo==2.0.1
open3d==0.19.0
open_clip_torch==3.0.0
openai==1.68.0
opencv-python==4.11.0.86
ordered-set==4.1.0
orjson==3.9.15
packaging==23.2
pandas==2.2.1
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
peft==0.8.2
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.2.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1715777629804/work
plotly==6.0.1
portalocker==2.4.0
preshed==3.0.10
promise==2.3
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1718047967974/work
protobuf==3.20.3
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1721585709575/work
py-cpuinfo==8.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.6.1
pycparser==2.22
pydantic==1.10.14
pydub==0.25.1
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
pyquaternion==0.9.9
pysbd==0.3.4
python-dateutil==2.8.2
python-multipart==0.0.9
python-slugify==6.1.2
pytorch-lightning==2.2.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq @ file:///croot/pyzmq_1705605076900/work
readchar==4.0.5
referencing==0.33.0
regex==2023.12.25
requests==2.32.4
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.7.0
rpds-py==0.18.0
rsa==4.9
runs==1.2.2
s3transfer==0.10.0
sacrebleu==2.1.0
safetensors==0.5.3
scikit-learn==1.4.1.post1
scipy==1.12.0
scispacy==0.5.5
semantic-version==2.10.0
sentence-transformers==4.1.0
sentencepiece==0.1.98
seqeval==1.2.2
shellingham==1.5.4
shtab==1.7.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smart-open==7.1.0
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.5.1
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
starlette==0.36.3
starsessions==1.3.0
stringcase==1.2.0
sympy==1.14.0
tabulate==0.8.10
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
tensorflow-addons==0.18.0
tensorflow-datasets==4.6.0
tensorflow-estimator==2.10.0
tensorflow-hub==0.12.0
tensorflow-metadata==1.9.0
tensorflow-model-optimization==0.7.2
text-unidecode==1.3
tf-slim==1.1.0
thinc==8.2.5
threadpoolctl==3.1.0
timm==1.0.19
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.0.1
torchaudio==2.0.2+cu117
torchmetrics==1.3.1
torchvision==0.15.2+cu117
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827245914/work
tqdm==4.67.1
traitlets==5.14.1
transformers==4.41.2
triton==2.0.0
trl==0.19.1
typeguard==4.4.4
typer==0.16.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.14.1
tyro==0.9.26
tzdata==2024.1
unsloth==2024.11.10
unsloth_zoo==2024.11.5
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.27.1
wasabi==1.1.3
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
weasel==0.4.1
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wrapt==1.17.2
xformers==0.0.31.post1
xmod==1.8.1
xxhash==3.5.0
yarl==1.9.4
zipp==3.21.0
